Added new header set for cleaning up Vk structs
Vulkan structures are C-based, thus no easy cleanup mechanism exists, so
this adds a function that attempts to do this, under the assumption that
any of the struct's pointer members with names of 'p[A-Z]' are owned by the
struct and can be safely
deleted.

Another members that are pointers that don't start with 'p[A-Z]' are
assumed to be externally managed and not deleted, nor any 'pointer to
pointers' type members.